app/requirements.txt
absl-py==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.34.8 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.34.8 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
flask==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
flatbuffers==23.5.26 ; python_version >= "3.11" and python_version < "4.0"
fonttools==4.47.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
matplotlib==3.8.2 ; python_version >= "3.11" and python_version < "4.0"
mediapipe==0.10.9 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.2 ; python_version >= "3.11" and python_version < "4.0"
opencv-contrib-python==4.8.1.78 ; python_version >= "3.11" and python_version < "4.0"
opencv-python==4.8.1.78 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
s3transfer==0.10.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sounddevice==0.4.6 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
